Move main error struct to separate file and rework app try_wait handling
